requirements.txt
# Don't forget to install requirements in ./third_party packages as well.
# See Dockerfile for example pip invocation.
djangorestframework==3.11.*
mod_wsgi-standalone
django-extensions
django-coreapi
pyyaml
